Restore no foreign code in shared
