+++ app/lib/database.ts
@@ -1,13 +1,14 @@
-import { DataAccessor } from '@liblab/data-access/dataAccessor';
 import { logger } from '~/utils/logger';
+import { DataSourcePluginManager } from '~/lib/plugins/data-access/data-access-plugin-manager';
 
 export async function executeQuery(connectionUrl: string, query: string, params?: string[]): Promise<any[]> {
-  const dataAccessor = DataAccessor.getAccessor(connectionUrl);
+  const dataAccessor = DataSourcePluginManager.getAccessor(connectionUrl);
 
   try {
     await dataAccessor.initialize(connectionUrl);
     dataAccessor.guardAgainstMaliciousQuery(query);
 
+  try {
     return await dataAccessor.executeQuery(query, params);
   } catch (e) {
     logger.error('Error executing query:', { error: e, query });
